Further refactoring of Items
