+++ src/blocks/Item/View.jsx
@@ -1,9 +1,21 @@
 import React from 'react';
 import Item from './Item';
 
+import { isArray } from 'lodash';
+import { serializeNodes } from 'volto-slate/editor/render';
+
+export const serializeText = (text) => {
+  return isArray(text) ? serializeNodes(text) : text;
+};
+
 const View = (props) => {
   const { data = {} } = props;
-  return <Item {...data} />;
+  return (
+    <Item
+      {...data}
+      description={data.description ? serializeText(data.description) : null}
+    />
+  );
 };
 
 export default View;
