+++ src/blocks/Listing/templates/Carousel.jsx
