+++ src/blocks/Listing/templates/NewsList.jsx
